Use native node.js apis instead of shelljs
I think shelljs makes a lot of sense if it the global version
is required and used massively throughout the program to make
it really look like a shell script. For our use case I think
it is better to use the native node.js apis 'fs' and 'child_process'
instead of installing a complete library to use it on four occasions
to save five characters. Also, even requiring the non-global
version of shelljs still monkeypatches the String prototype with
a #to convenience method (basically a wrapper around
fs.writeFileSync) and probably also other things that I’m not
aware of. And I think either you buy into the full global thing
or not, but that subtle monkeypatching seems dangerous, because
it is monkeypatching and monkeypatching is dangerous.